Displaying user balance

+++ payapp/models.py
@@ -11,7 +11,7 @@ class UserProfile(models.Model):
         ('EUR', 'EUR'),
     ]
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='payapp_profile')
-    bal = models.IntegerField(default=0)
+    bal = models.FloatField(default=0)
     currency = models.CharField(max_length=3, choices=CURRENCY_CHOICES, default='GBP')
 
     def __str__(self):
